requirements.txt
streamlit==1.37.0
google-genai==0.3.0
pillow==10.3.0
python-dotenv==1.0.0
requests==2.32.4
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
flake8==6.1.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
Pillow==10.3.0
streamlit==1.37.0
